backend/requirements.txt
Django==2.2.3
psycopg2-binary==2.8.3
djangorestframework
django-filter
djangorestframework-jwt
django-cors-headers